Kernel: Move LockMode into Locking/
